refactor structure of LeftSideBar. replace LeftSideBar by UserInfo, UserInfo and ChatList(listUser)
