+++ src/components/homePage.jsx
@@ -1,9 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import withAuth from './withAuth'; // Import HOC withAuth
 import '../assets/style/homePage.css'; // Import CSS file
-import LeftSideBar from "./leftSideBar/LeftSideBar";
+// import LeftSideBar from "./leftSideBar/LeftSideBar";
 import MainChat from "./mainChat/MainChat";
 import RightSideBar from "./rightSideBar/RightSideBar";
+import UserInfo from "./leftSideBar/UserInfo";
+import ChatList from "./leftSideBar/ChatList";
+import SearchBox from "./leftSideBar/searchBox";
 
 // Component của trang HomePage
 const HomePage = () => {
@@ -109,11 +112,17 @@ const HomePage = () => {
 
     return (
         <div className="App">
-            <LeftSideBar
-                handleCreateRoom={handleCreateRoom}
-                handleJoinRoom={handleJoinRoom}/>
-            <MainChat />
-            <RightSideBar />
+            <div className='leftSideBar'>
+                <UserInfo/>
+                <div className='chatList'>
+                    <UserInfo
+                        handleCreateRoom={handleCreateRoom}
+                        handleJoinRoom={handleJoinRoom}/>
+                    <ChatList/>
+                </div>
+            </div>
+            <MainChat/>
+            <RightSideBar/>
         </div>
     );
 };
